requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.8
django-audit-log==0.7.0
django-celery==3.1.17
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
MySQL-python==1.2.5
python-ldap==2.4.27
pytz==2016.6.1
redis==2.10.5
